Fixed some 64-bit/32-bit casting issues with `unsigned int` and `size_t`.
- Changed some variables to `size_t`.
- Added explicit casts in other places.

This fixes several of the following MSVC warnings:
- warning C4267: '=': conversion from 'size_t' to 'unsigned int' or 'DWORD', possible loss of data
- warning C4838: conversion from 'DWORD' to 'int' requires a narrowing conversion

+++ src/libromdata/Audio/SNDH.cpp
@@ -663,7 +663,7 @@ SNDH::SNDH(IRpFile *file)
 	// Check if this file is supported.
 	DetectInfo info;
 	info.header.addr = 0;
-	info.header.size = size;
+	info.header.size = static_cast<uint32_t>(size);
 	info.header.pData = buf;
 	info.ext = nullptr;	// Not needed for SNDH.
 	info.szFile = 0;	// Not needed for SNDH.
